feat(l1): added validation before removedb (#2419)
Clone of #2125

**Motivation**

- Adds a layer of validation before executing the `removedb` command
- Add `force` flag to delete without confirmation
<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Asks the user for confirmation before executing the `removedb` command
and deleting the database


To test:
```
cargo run removedb
```
or 
``` 
cargo run removedb --force
```

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #602

---------

+++ cmd/ethrex/cli.rs
@@ -1,5 +1,6 @@
 use std::{
     fs::{metadata, read_dir},
+    io::{self, Write},
     path::Path,
 };
 
@@ -51,6 +52,14 @@ pub struct Options {
         help_heading = "Node options"
     )]
     pub datadir: String,
+    #[arg(
+        long = "force", 
+        help = "Force remove the database",
+        long_help = "Delete the database without confirmation.",
+        action = clap::ArgAction::SetTrue,
+        help_heading = "Node options"
+    )]
+    pub force: bool,
     #[arg(long = "syncmode", default_value = "full", value_name = "SYNC_MODE", value_parser = utils::parse_sync_mode, help = "The way in which the node will sync its state.", long_help = "Can be either \"full\" or \"snap\" with \"full\" as default value.", help_heading = "P2P options")]
     pub syncmode: SyncMode,
     #[arg(
@@ -182,6 +191,7 @@ impl Default for Options {
             metrics_port: Default::default(),
             dev: Default::default(),
             evm: Default::default(),
+            force: false,
         }
     }
 }
@@ -193,6 +203,8 @@ pub enum Subcommand {
     RemoveDB {
         #[arg(long = "datadir", value_name = "DATABASE_DIRECTORY", default_value = DEFAULT_DATADIR, required = false)]
         datadir: String,
+        #[clap(long = "force", help = "Force remove the database without confirmation", action = clap::ArgAction::SetTrue)]
+        force: bool,
     },
     #[command(name = "import", about = "Import blocks to the database")]
     Import {
@@ -213,14 +225,15 @@ pub enum Subcommand {
 impl Subcommand {
     pub async fn run(self, opts: &Options) -> eyre::Result<()> {
         match self {
-            Subcommand::RemoveDB { datadir } => {
-                remove_db(&datadir);
+            Subcommand::RemoveDB { datadir, force } => {
+                remove_db(&datadir, force);
             }
             Subcommand::Import { path, removedb } => {
                 if removedb {
                     Box::pin(async {
                         Self::RemoveDB {
                             datadir: opts.datadir.clone(),
+                            force: opts.force,
                         }
                         .run(opts)
                         .await
@@ -242,16 +255,30 @@ impl Subcommand {
     }
 }
 
-pub fn remove_db(datadir: &str) {
+pub fn remove_db(datadir: &str, force: bool) {
     let data_dir = set_datadir(datadir);
-
     let path = Path::new(&data_dir);
 
     if path.exists() {
-        std::fs::remove_dir_all(path).expect("Failed to remove data directory");
-        info!("Successfully removed database at {data_dir}");
+        if force {
+            std::fs::remove_dir_all(path).expect("Failed to remove data directory");
+            println!("Database removed successfully.");
+        } else {
+            print!("Are you sure you want to remove the database? (y/n): ");
+            io::stdout().flush().unwrap();
+
+            let mut input = String::new();
+            io::stdin().read_line(&mut input).unwrap();
+
+            if input.trim().eq_ignore_ascii_case("y") {
+                std::fs::remove_dir_all(path).expect("Failed to remove data directory");
+                println!("Database removed successfully.");
+            } else {
+                println!("Operation canceled.");
+            }
+        }
     } else {
-        warn!("Data directory does not exist: {data_dir}");
+        warn!("Data directory does not exist: {}", data_dir);
     }
 }
